changes to big decimal: accept double and int

+++ common/src/main/java/edu/ast/expressions/LiteralNumberNode.java
@@ -3,8 +3,9 @@ package edu.ast.expressions;
 import edu.LexicalRange;
 import edu.ast.AstVisitor;
 import edu.ast.interfaces.LiteralNode;
+import java.math.BigDecimal;
 
-public record LiteralNumberNode(LexicalRange start, LexicalRange end, Double value)
+public record LiteralNumberNode(LexicalRange start, LexicalRange end, BigDecimal value)
     implements LiteralNode {
 
   @Override
